- Convert economy from long to BigInteger - Change StringUtils.numberFormat to support BigInteger, now with an iterative implementation - Add new suffixes for bigger numbers - Remove already default null fallback parameter in event.getOption calls - Added database codecs for BigDecimal (store full decimal instead of Decimal128), BigInteger, Color, and Newsitem - Added GuildData.getOrCreateGuildData and replaced all other GuildData instantiations with that method - Fixed "next birthday" age being previous birthday age - No longer use deprecated getWhen method in ChangelogFetcher - Fixed documentation of CrashCommand.Game.getMultiplier not reflecting the implementation - /crime demotion message: "demoted by -1 level(s)" -> "demoted by 1 level" - Update gradle version to 8.12.1 - Fixed a few typo'd "baos"es - In NotifierCommand, "setup" -> "set up" - Make RewardCommand more DRY and make it easier to add new types of rewards - Compact /tag list - In ThreadManager, created setupThread to use some code twice instead of duplicating the code - Fix TimeUtils.calculateMaxBirthYear - In TurtyBot, no longer use deprecated GatewayIntent.GUILD_EMOJIS_AND_STICKERS

+++ src/main/java/dev/darealturtywurty/superturtybot/modules/economy/Payment.java
@@ -4,10 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigInteger;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class Payment {
     private long time;
-    private long amount;
+    private BigInteger amount;
 }
